Adding better debug for failure to cleanup tmpFile
Turns out that windows falls back to trying to create a file with the
equivalent of the O_CREAT if it fails to stat a file the conventional way. This
manifests as a CreateFile failure if the file exists (which we see above),
which comes back through the os/syscall layers undecorated with any more
information and all we see is CreateFile: Access is denied. Thanks Windows.

[#159728888]

Co-authored-by: Tom Godkin <tgodkin@pivotal.io>